requirements.txt
pytube
timm
transformers
torchvision



